pyproject.toml
[tool.poetry]
name = "chirpage"
version = "5.2.0"
description = ""
authors = ["OverflowCat <overflowcat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
html-telegraph-poster = "^0.5.0"
gspread = "^5.12.3"
python-dotenv = "^1.0.0"
python-telegram-bot = { extras = ["all"], version = "^20.7" }
Pillow = "^9.5.0"
python-bidi = "^0.4.2"
pysos = "^1.3.0"
schedule = "^1.2.1"
tinydb = "^4.8.0"
PyMySQL = "^1.1.0"
tweepy = "^4.14.0"
textile = "^4.0.2"
termcolor = "^2.4.0"
openai = "^0.27.10"
duckduckgo-search = "^2.9.5"
cachetools = "^5.3.2"
result = "^0.9.0"
rich = "^13.7.0"
async-cache = "^1.1.1"
peewee = "^3.17.0"
poe-api = "^0.4.17"
selenium = "^4.16.0"
google-generativeai = "^0.3.1"
ruff = "^0.1.8"


[project]
authors = [{ name = "OverflowCat", email = "overflowcat@gmail.com" }]
requires-python = ">=3.10,<4.0"
dependencies = [
    "html-telegraph-poster<1.0.0,>=0.2.0",
    "gspread<6.0.0,>=4.0.0",
    "python-dotenv<1.0.0,>=0.14.0",
    "python-telegram-bot<21.0,>=20.0",
    "Pillow<10.0.0,>=9.4.0",
    "python-bidi<1.0.0,>=0.4.2",
    "pysos<2.0.0,>=1.2.1",
    "schedule<1.0.0,>=0.6.0",
    "tinydb<5.0.0,>=4.7.0",
    "PyMySQL<2.0.0,>=1.0.2",
    "tweepy<5.0.0,>=4.13.0",
    "textile<5.0.0,>=4.0.2",
    "cachetools<5.0.0,>=4.2.2",
    "termcolor>=2.2.0",
    "result>=0.9.0",
    "openai>=0.27.2",
    "poe-api>=0.4.3",
]
name = "chirpage"
version = "0.1.0"
description = ""

[project.urls]
